Refactor and improve documentation in msp_utilities.py
The function documentations in msp_utilities.py script have been streamlined and detailed for better clarity. The focus was on improving the precision of return types, simplifying the explanations and making them directly correspond to the function's processing. This will ease understanding of the role and operation of various functions dealing with spectra processing.