Improve DataFrame formatting and peak identification in converters.py
In the DataFrame formatting function in converters.py, the row formatting process has been enhanced. Also, added a new function to check if a value represents a peak. Depending on this check, column names may be renamed to 'PEAK_LIST'. Furthermore, the opening mode of files in MAIN.py script was changed from "append" to "write".