Add function to remove peaks above precursor MZ in spectrum
This commit adds a function called "remove_peaks_above_precursor_mz" that removes the peaks above the precursor MZ in the given spectrum. This is implemented in the "matchms_processing.py" script alongside modifications in the "filters.py". Functions for the conversion of data types and reformatting of the peak list are also included. The efficiency of spectrum processing is expected to be improved with these changes.