Add test.py to generate and process spectrums
The new file, test.py, in the '/FragHub/scripts/Drafts' directory, generates a list of spectrums from '.msp' files and processes them. It ensures only non-empty spectrums are used for processing and saves each processed spectrum as an individual Excel file in the OUTPUT directory. The total execution time of the script is also displayed.