Add 'keep_mz_in_range' method in Filters_NEW.py
A new function 'keep_mz_in_range' has been added to filter the peak dataframe to keep only the rows where the 'mz' column value is within the range of 50 and 2000. This filtering is now included in the 'apply_filters' method workflow to ensure the 'mz' values are always within the specified range.