Add high peak check in peak dataframe filtering
A function 'check_minimum_of_high_peaks_required' has been added to check if a peak dataframe has a sufficient number of high peaks. It filters the dataframe based on a given intensity percentage and requires a minimum number of peaks. This check is now included in the 'apply_filters' function.