Add 'normalize_predicted' function to values_normalizer.py
The values_normalizer.py script has been updated to include the 'normalize_predicted' function. This function sets the 'PREDICTED' key to 'true' in the metadata_dict if the comment field matches our In_Silico_pattern or the 'PREDICTED' value is 'true'. If not, it sets it to 'false'. This change assists in normalizing predicted values in the metadata dictionary.