Update 'NUM PEAKS' assignment in msp_normalizer script
Modified the method for calculating the 'NUM PEAKS' in the metadata of the msp_normalizer script. Previously, it was calculated by the length of the peak_list, it is now updated to be calculated as the number of new lines in the peak_list plus one, making it more accurate.