Modify path of update.json in update.py
The path to the 'update.json' file has been adjusted from the relative path './datas/update.json' to '../datas/update.json' in the 'update.py' script. This change updates the file handling operations to refer to correct directory structure. The change was implemented in all instances where the file is opened for both reading and writing purposes.