Add filename check in values normalizer
The "values_normalizer.py" script has been updated to include a check of the file name. This addition involves the creation of a function "in_filename" which checks if the string "MSMS_Public" is within the filename and if it matches a certain pattern. The "normalize_predicted" function was also modified to incorporate the filename checking.