Reposition sleep timers in write scripts
The sleep timer of 0.1 second has been redistributed in write_msp and write_csv functions in the script writers.py. Instead of a sleep timer at the start of the functions, it's now called after writing and deleting each file. This should aid in appropriate memory allocation and management during cyclic operations.