Comment out update check in MAIN.py
The update check function call in MAIN.py has been commented out. The check for update function was previously used to update `spectrum_list`, but is currently disabled, likely for testing or troubleshooting purposes. The consequences of this modification on the rest of the code are not evident from the diff.