Add filename parameter to fraghubid processing function
The genrate_fraghubid_processing function in the fraghubid_generator.py script has been updated to accept a filename parameter. This allows for a more specific description message during the fraghubid generation process. A filename is now also passed when the genrate_fraghubid_processing function is called in the loop over files ending with ".msp".