Adjust text alignment for progress messages
The main change includes a modification in the HMDB_completion.py script where the progress text messages have been centered. More specifically, the update allows progress message and spectra completion description to be tidily aligned to the right side of the screen, instead of being spread across the full breadth of the console window.