Refactor json file search and concatenation in converters.py
This commit changes the method for finding and concatenating json files. Instead of using os.listdir and accepting directories as parameters, the function now accepts a list of file paths, obtained through os.walk. This allows for a more versatile use of the function, being capable of concatenating files from multiple directories.