Refactor XML file search and concatenation in converters.py
This commit modifies the method for locating and concatenating XML files. Previously, os.listdir was used and directories were taken as parameters. Now, the function accepts a list of file paths, gathered via os.walk. This new approach enhances the versatility of the function, enabling it to concatenate files from various directories.