Change directory and file type in MAIN.py script
This commit changes the directory the script walks through from 'msp_dir' to 'json_dir', and also alters the type of file it looks for from ".msp" to ".json". A small delay is also added to the cleaning process. Moreover, the 'CONCATENATED_SPECTRUMS_RESULTS' array extension and the deletion of 'spectrum_list' have been commented out for future consideration.