Update JSON file used in get_first_n_items function
The JSON file used as an argument in the get_first_n_items function has been updated in the test.py script. Previously, it was 'MSP_converted.json', but it has been changed to 'ALL_GNPS.json'. This change aligns the script with the current project files and ensures correct and updated data outputs.