Add JSON processing to fraghubid_generator
Integrated JSON file support in the fraghubid_generator script. The script has been updated to accept JSON input files besides MSP, XML, and CSV. This has entailed the introduction of a new condition for file path assignment and a corresponding update to the 'if' clause in the generate_fraghub_id function. Now, the script can process a wider range of file formats.