Update large JSON processing in test.py script
The code has been significantly updated to include a function for processing large JSON files, named `process_large_json`. This function now takes an input and output file name as well as a specified number of records to process from the input file. All records retrieved are written into the output file with pretty formatted JSON. This enhancement allows the script to process a specified number of records from a large JSON file.