Update msp_normalizer.py to replace JSON with eval and str functions
The usage of json.loads and json.dumps functions in msp_normalizer.py has been replaced with eval and str functions respectively. This change was made to avoid the unnecessary overhead of converting to and from JSON. The condition to return empty string for zero-sized peak list has been moved down to avoid executing unnecessary code.