Set default "filename" in spectrum dictionary
This change eliminates the condition checking for the presence of "filename" key in the spectrum dictionary. Instead, the "filename" is now added directly from the json_file_path's basename, simplifying the code and taking advantage of the os.path.basename()'s output, regardless of whether the path is a directory or a file.