Update installation and usage instructions in README
The changes made in README.md file simplify the installation and usage instructions. The mention of HMDB xml spectrums processing has been removed in the instructions, making it straightforward for the user to run scripts according to their OS. The instructions have been improved for better understanding and ease of use.