Update main script to conditionally perform tasks based on parameters
This update introduces condition checks for 'msp', 'csv', and 'json' from the parameters dictionary. As a result, the tasks of converting CSV to MSP, writing CSV, MSP, and JSON outputs will only be carried out if the corresponding parameter value is set to 1.0, providing flexibility in script execution based on user configuration.