Add Generate_csv_from_original_files script to FragHub
A new script, named 'Generate_csv_from_original_files' was added into the FragHub project under TOOLS directory. This script will be instrumental in generating csv files directly from the original ones. The addition of this script aims to facilitate data handling and improve the project organization for the user.