Add file existence check before reading in duplicatas_remover.py
This commit improves fault tolerance in 'duplicatas_remover.py' by checking the existence of a file before attempting to read from it. This avoids errors when the respective file does not exist and instead initializes an empty DataFrame in such cases. Additionally, the 'os' module has been imported to facilitate this check.