Change return value from empty dict to None in msp_normalizer.py
In msp_normalizer.py, two return statements were modified to return None instead of an empty dictionary when float_precursor_mz is less than or equal to 0, or peak_list_np's size is 0. This change could better handle these edge cases and improve the readability and understandability of the code.