Add new code cells to display total spectra and file counts
This commit involves the addition of new code cells to the "upset_msp.ipynb" and "Generate_csv_from_original_files.ipynb" notebooks. These new cells output the total number of spectrums for graphs and the sum of various file types, respectively. This allows for better tracking and overview of data quantity. Additionally, the change of the path to the "original_db_path" variable in "Generate_csv_from_original_files.ipynb" is included in this commit.