Update and refactor data cleaning in multiple notebooks
Refactored and upgraded data cleaning operations in the 'MSP_INCHIKEYs_excel.ipynb' and 'add_classyfire_to_msp.ipynb' notebooks. Changes include the addition of execution times, conversion of dataframe columns to the appropriate data type, and removal of duplicates. This ensures a more accurate data analysis and improves the efficiency of the code.