Add Classyfire classes and subclasses to MSP data
Created a new Jupyter notebook that takes the results from Classyfire and appends them to the MassBank (MSP) data. The additions include the Classyfire kingdom, superclass, class and subclass. In this update, the previous null values in the Classyfire columns in the MSP data were replaced. Furthermore, extra columns that surfaced during the process, which were not needed, were removed.