Simplify color generation in Python script
The generate_colors_dict.py script in the TOOLS directory is refactored to simplify the color generation process. It now uses the inbuilt 'colorsys' and 'colors' modules from matplotlib library for generating a dictionary of unique RGB colors, instead of the previous complex method of color conversion. The list of categories has also been modified and cleaned up.