Add exact molecule weight calculation and spectrum cleaning enhancements
Updated the `spectrum_cleaning` function in the `msp_normalizer.py` script, enabling it to handle precursor mass recalculation when necessary. This improvement also includes the implementation of a new function `precursor_mz_re_calculation` to calculate the precursor m/z for a given molecular structure and a mass difference. The dictionary `adduct_massdiff_dict` was added to store data from the `adduct_to_convert.csv` file. These updates make the script more flexible and comprehensive in dealing with spectrum data.