Remove assignment of "UNKNOWN" to IONIZATION
The change in the code removes the else branch that previously assigned the value "UNKNOWN" to the IONIZATION key in the metadata_dict dictionary. This adjustment means that if ionization_mode_in_INSTRUMENTTYPE is not "ACPI", the "IONIZATION" field will not be set, removing potential confusion or incorrect data marking.