Update `apply_filters` method parameters
Removed `parameters_dict` import from the peaks_filters/filters.py file and added it as a parameter to the `apply_filters` method. This change was also reflected in the msp_normalizer.py file where `apply_filters` method is invoked. This allows for better code modularity and prevents repeated import statements.