Add new permutation.ipynb tool to FragHub
The new tool, permutation.ipynb, has been added to the TOOLS directory of FragHub. This new tool generates permutations of a dataframe by replacing various combinations of columns in the dataframe with a placeholder value, 'UNKNOWN'. It then applies a variety of filtering rules to prune unwanted permutations and exports the resulting dataframe as an Excel file.