Update and expand functionality in FragHub TOOLS
Implemented the addition of new tools to the FragHub project's toolkit, which include the complete_tree_solution.ipynb and check_LC_GC.ipynb tools. Also improved the instruments_graph.ipynb tool with the inclusion of the 'SOLUTION' column in the tree structure. Furthermore, the added tools provide, respectively, a formatted solution for every possible instrument setup given the resolution and ionisation type, and the validation of potential IONISATION values for each SPECTRUM_TYPE.