Update data source and execution count in decision tree notebook
This commit updates the source data file used in the check_decision_tree notebook under the FragHub toolset. The data source from "Shimadzu" sheet is switched to the "TOTAL" sheet. Additionally, the execution count is also updated across "code" cells. The change in the data source is expected to impact the output generated by the notebook.