Implement separate cleaning functions for instrument data
Refactored the normalization of instrument and instrument-type fields in the normalize_instruments_and_resolution module. A separate function was created to clean both instrument and instrument type, which now replaces the instant modification in clean_spectrum_instrument_info function. This change not only enhances the legibility and maintainability of the code but also improves modularity.