Update check_for_update_processing function in update.py
Improvements were made to the check_for_update_processing function in the update.py script. The returned object was changed from a list of updated spectrums to a tuple containing the updated spectrum list, a flag indicating if an update was found, and a first run flag. Moreover, a new parameter, profile_name, was added to specify the profile to use for checking updates.