Refine remove_peak_above_precursormz function documentation
The detail of revise focuses on function `remove_peak_above_precursormz`, in the `peaks_filters` script. The function documentation is further clarified with more explicit details regarding the parameters and the functionality of the filtering criterion. Also, a bit more error handling is added for when the input `precursormz` is not a float.